requirements.txt
discord.py
numpy
pynacl
pyttsx3
